app/requirements.txt
Flask
Flask-Login
Flask-Bootstrap4
Flask-Wtf
uwsgi
psycopg2